+++ transport/mtproxy.go
@@ -5,7 +5,7 @@ import "github.com/gotd/td/internal/proto/codec"
 // MTProxy creates MTProxy obfuscated transport.
 //
 // See https://core.telegram.org/mtproto/mtproto-transports#transport-obfuscation
-func MTProxy(d Dialer, dc int16, secret []byte) *Transport {
+func MTProxy(d Dialer, dc int, secret []byte) *Transport {
 	return NewTransport(d, func() Codec {
 		return &codec.MTProxyObfuscated2{
 			Codec:  codec.PaddedIntermediate{},
